First few Arduino pages set up!
Going to do the General Electronics pages, followed by the Product Reviews.

Images are done so that the browser will load the .webp files, if a visitor's browser supports displaying WebP images. If the browser does not support WebP, then it would load the .png and .jpeg/.jpg files.

Because WebP files are smaller, and if the browser supports WebP, *only* .webp files will be transferred, saving the visitor time and bandwidth.

There is more HTML code to write that will get transferred over, but more individual bytes for each character in the HTML code on each page is better than more *kilo*bytes for each image that gets exchanged for less HTML code that would be just <img> tags all over with no <picture> tag as its parent.