Finished first two pages of General Electronics
Also added the two new pages you made to be ready to finish later.

I realised that I might have to surround all <picture> blocks and clusters with a <div> to allow flexbox for responsiveness; the <div> would be the flexbox parent, while the <picture>s are the flexbox children.

Added comments in the .yml files to help follow the scheme when adding new pages in the future.