Two more General Electronics pages done!
Made a table in 'Project 004: Cadsoft Eagle to Gerber Conversion Guidelines' to better show the file formats' listed specifications.
Proofreading and appropriate `alt=""` attributes will have to come later since writing the text for each image takes a lot of time that could be used for transferring the content and styling sooner.