Product Review 005 for ElecFreaks Micro:bit Tinker Kit finished!
Now I can begin styling.

---

Four things about this page:

1: I decided to name the images as how they really are, instead of something like "19864153.jpg" as they're named straight from Weebly. I thought this would make managing the assets in the future easier.

2: Original links from the Weebly site are now HTTPS links, since HTTP-only links currently take the user to ElecFreaks' homepage.

3: All links that link to ElecFreaks' site (and from hereon for all external sites) and open in a new tab will have this added: rel=”noopener noreferrer”
This is apparently a bit of protection against "tabnabbing", which is malicious pages opening in a new tab (because of target="_blank") taking control of the originating tab's content to trick users when they go back to that originating tab.
External sites could have been compromised and loaded with the tabnabbing attack. It may not be obvious *if* the site has been compromised at all.
https://mathiasbynens.github.io/rel-noopener/

4: Relating to Point 2, the ElecFreaks link for the Tower Pro SG-90 Mini Servo Motor is broken because ElecFreaks announced on their site that they're now only selling Octopus and Micro:bit products, even if their products includes non-Octopus or -Micro:bit products. The servo motor, even though it's included in the kit, is not from Octopus nor Micro:bit, so there no longer won't be any separate listing on their site. There is also no kind of archive of that original page at all.
So, the header for the servo motor won't have any link right now, though there are non-ElecFreaks listings of that same motor. Not sure if you prefer a link in its place or no link.