Some styling implemented!
I spent so much time on the home page that I'm starting to consider using the current styles.css file as a .css file *just* for the homepage, because some declarations in styles.css appear only on the homepage; it would be a waste of user bandwidth to download CSS styling that won't be used on any non-homepage page they will be on.

I also ended up going with a mobile-first approach, since sites generally are always getting mobile visitors than desktop visitors. Media queries would be needed to switch the layout to a more desktop-friendly layout.
The homepage right now looks better as it is right now with this commit, even on the desktop, but some things like fonts, text positioning (especially when viewing on mobile), and the colour-changing transition on the button links are still left.