Text sizing and margins/paddings, with added responsiveness
Line spacing to be more on par with sites' mobile designs to allow more text without looking too cramped and too close to the display's edges.

Fixed the flexboxes a bit so they won't be so funky with the newer margin/padding changes.

Mobile view goes up to 800 pixels, which then transitions to a more desktop-like viewport. 

`.hero-text * {}` needs its own line-height because `* {}` was changed for everything in the page. That ended up changing the line spacing for the text in the hero area, and I had to get the text to how I had it when `.hero-text * {}` was as 1.8em.