Successfully added transparent overlay on top of banner image
- `background-repeat` does nothing unless the image is not big enough. Probably best to keep it there
- Gradients with the hero image URL in a variable seems to be the best way to get a transparent overlay in one selector, `.hero`, instead of two selectors, `.hero` and `.hero-text`
----`background` declaration in `.hero-text` was then removed.
- `background-size` and `background-position` have to be below `background` or else they will get ignored

Another reason for choosing this option is so you can easily change the hero image without affecting the CSS for the overlay. Simply change the URL in the `<header>` tag and the new image will replace the old one. Or replace the current one with a new one with the same file name. Whichever is done.
Inline styling in the HTML with `background-image` won't work since the `background` solution requires an image URL, which it is getting from the HTML as a CSS variable.
There's also an explanation for that area in `.hero` for if a new hero image has different dimensions from the current one.

Sources:
https://web.archive.org/web/20200326235050/http://www.selbekk.io/blog/2019/06/how-to-overlay-your-background-images/
https://web.archive.org/web/20200326235242/https://webdevetc.com/blog/how-to-add-a-gradient-overlay-to-a-background-image-using-just-css-and-html