Fixed navbar bottom border floating out
No longer an "outline"-border, but now a border itself.
Borders add to a content's box size. So wherever the box height is hard-defined, the non-border content gets squeezed.
To counter that, the total height of the navbar is now the original height, *plus* the height of the original border. The end result is the same visual output, but done with a different method that won't make the border floating out of the navbar area.