Navbar height adjustment
Navbar on both <800px and >800px now at 5rem and 7rem, respectively.
Had to add flex ruling to <nav> so that all items would still be centred; simply adjusting the navbar height leaves the logo uncentred. 

Line height shortened for mobile view so that the hover effect wouldn't flow outside of the navbar space, and to make the menu items more appropriate on mobile view.