Clicking on menu items no longer slightly moves other items
`:focus`'s borders add to the total amount of content in the group and end up moving the menu items' positions very slightly. It doesn't look so great.
Using outlines doesn't have that issue, so it would be best to use them.

Visual output when clicking on the items is the same, but all items being static when one of them is clicked on is good polish to have.