Kind of a big rewrite of the code for the content
When writing the contact.html page, I noticed that the .content div in /_layouts/default.html wouldn't do well with pages not written like the Homepage. So, I had to use `main > *` that I had then recently discovered, as well as the other `main > ` declarations.

Doing all that meant that I had to change the margins for .hero and .section-border at the same time to get the same visual end result, while giving *all* pages consistent margin and padding styling.

It doesn't look like a lot changed, but it took so much trial-and-error-ing to get to this. At this point, styling the form should be easier now.