Removed footer links and files/code referencing those links
Removed footer links since the menu bar will be sticky and constantly available. If footer links are going to be gone, so would the .yml file for each individual link as well as the styling; best to get everything about it removed and not leave unneeded stuff behind to be removed in a later commit.