More-specific targeting of children under <main>
Honestly, I can't completely remember what this was for in styles.css, but I know that I need it like this. I think it was to make the widths of non-.hero and non-.section-border children to be more in line with how I will be making the contact form's. Reinforcing the uniformity of all content.

This is what I get for not writing detailed comments for me to reference at a later date.