New layout for homepage to work around <hr> issue
The <hr> bars on the the non-homepage pages weren't confined like in the homepage, thus making them span beyond the content's width, even touching the viewport widths when viewing on mobile.

The best fix I could make for those pages was to add a <div> under <main> to contain the content. Because doing that would be a pain to do for all pages, as well as be a bit error-prone, I decided to make the current default.html layout have that <div>, while the homepage gets its own layout *without* that <div>.

"Default" also happens to be more semantic to how the pages are set up, so it's kind of like a plus.