Further styling on contact form
The form now looks prettier!

The field for the names align in a row on a large viewport, but wrap to 
their own lines on mobile. Thought it would be best since text would be 
too cramped if both fields stayed aligned in a row.
Putting the fields for the names in a div was pretty much the only way 
to easily do it.

<label>s have bottom spacing to give each section some room.

Button given a width so that it don't span the whole form width.