New hero image
New hero image is currently a bit bloated to be first loaded when a 
visitor first loads the homepage. Will need to figure out how to 
decrease the size without the browser not seeing the image is something 
it can't or just won't load, even though the image is referenced 
correctly.

Co-Authored-By: null <Electorials@users.noreply.github.com>