Hero rewrite for image blur
Reference/help: 
https://www.w3schools.com/howto/howto_css_blurred_background.asp

Image is similar to how it was in the <header> tag, but now in a <div> 
that is a sibling to
another <div> that holds the text.

The `--image-height: 35rem` line keeps the size of the .hero section and 
the actual image in sync
in case you wish to change the height of the whole .hero section.

The `position: absolute` is what gets the text to be overlayed on top of 
the image.
`top: 14rem` for vertical positioning to still look fine on mobile and 
on desktop.