Grid implemented on Projects and Reviews pages
Grid layout implemented with values from the pages' YML files in /_data/grid-items/.

Page entries now inside a group named "items". This is for a bit of organisation and letting the grid layout parts reside in the same file without possibly inadvertently messing something else up.

I couldn't figure out if using one layout file for all Projects and Reviews pages was possible, so there's now a separate layout page for the corresponding pages.

Lots of the removals in styles-grid.css was from copying what was made to styles.css. They weren't needed, so removing them wouldn't hurt.

Added a README in /_data/grid-items/ that explains how to modify the values for the grid layout in the /_data/grid-items YML files.